ingest_service/requirements.txt
fastapi
uvicorn
influxdb-client
python-jose[cryptography]
pandas
python-multipart
pytest
pytest-cov
